Support minitest-given use in Rails tests
This change is very basic: if ActiveSupport::TestCase is defined, then
require minitest-given under an artificial circumstance in which 
Minitest::Spec is aliased to ActiveSupport::TestCase (so that TestCase
absorbs all of the monkey-patching normally applied to Minitest::Spec).

This will be a breaking change to anyone currently using Minitest::Spec
directly from a test suite that loads Rails, but it's not clear to me
that this is a common (if possible?) use case. If that use case matters,
it may be necessary to force-load all the files again after un-aliasing
the Minitest::Spec constant.

lib/minitest-given was changed so that Bundler's auto-require would pick
up the same intelligence that `require 'minitest/given'` will now have.