Isolate branching & reworking the metadata spec
Having the if-else branch wrap the entire describe confused me, because
I couldn't tell whether the metadata support we offer differs 
substantially between RSpec 2 & 3 (it conflated the fact that RSpec 3
added default-true metadata, which 2 lacks with the fact the `example`
DSL was removed from RSpec 3)

I decided to make the default-to-true metadata example more explicit, as
well, including a slightly more rugged rspec_3_or_later? helper available
to the rest of the unit test sutie.